requirements.txt
pytest~=8.2.2
pytest-playwright~=0.5.0
pytest-html~=4.1.1
allure-pytest~=2.13.5
pytest-xdist~=3.6.1
Faker~=26.0.0
python-dotenv~=1.0.1 